+++ src/cashier_frontend_new/src/modules/creationLink/state/linkCreationStatesV3/preview.spec.ts
@@ -1,8 +1,5 @@
 import { LockStateV3 } from "$modules/creationLink/state/linkCreationStatesV3/lock";
-import {
-  gateDraftToGateKey,
-  PreviewStateV3,
-} from "$modules/creationLink/state/linkCreationStatesV3/preview";
+import { PreviewStateV3 } from "$modules/creationLink/state/linkCreationStatesV3/preview";
 import { LinkCreatedStateV3 } from "$modules/creationLink/state/linkCreationStatesV3/created";
 import type { LinkCreationStoreV3 } from "$modules/creationLink/state/linkCreationStoreV3.svelte";
 import type { CreateLinkResponseV3 } from "$modules/creationLink/types/dto/create_link_v3";
@@ -127,60 +124,6 @@ function makeStore(options?: {
   };
   return store as unknown as LinkCreationStoreV3;
 }
-
-describe("gateDraftToGateKey", () => {
-  it.each([
-    [
-      "password",
-      { type: GateType.PASSWORD, password: "secret" },
-      { Password: "secret" },
-    ],
-    [
-      "x following",
-      {
-        type: GateType.X_FOLLOWING,
-        targetHandle: "cashierapp",
-        rewardAccount: "reward-account",
-      },
-      { XFollowing: "cashierapp" },
-    ],
-    [
-      "x owned account",
-      { type: GateType.X_OWNED_ACCOUNT, targetHandle: "cashierapp" },
-      { XOwnedAccount: "cashierapp" },
-    ],
-    [
-      "x liked post",
-      { type: GateType.X_LIKED_POST, tweetUrl: "https://x.com/a/status/1" },
-      { XLikedPost: "https://x.com/a/status/1" },
-    ],
-    [
-      "x retweeted post",
-      { type: GateType.X_RETWEETED_POST, tweetUrl: "https://x.com/a/status/2" },
-      { XRetweetedPost: "https://x.com/a/status/2" },
-    ],
-    [
-      "otp email",
-      { type: GateType.OTP_EMAIL, email: "user@example.com" },
-      { OTPEmail: "user@example.com" },
-    ],
-    [
-      "otp sms",
-      {
-        type: GateType.OTP_SMS,
-        phone: "+15555550123",
-        digits: "5555550123",
-        countryCode: "US",
-      },
-      { OTPSms: "+15555550123" },
-    ],
-  ] satisfies Array<[string, GateDraft, unknown]>)(
-    "it_should_map_%s_gate_draft_to_backend_gate_key",
-    (_, gateDraft, gateKey) => {
-      expect(gateDraftToGateKey(gateDraft)).toEqual(gateKey);
-    },
-  );
-});
 
 describe("PreviewStateV3", () => {
   beforeEach(() => {
