+++ src/cashier_frontend_new/src/modules/creationLink/state/linkCreationStatesV3/preview.ts
@@ -5,34 +5,14 @@ import type { LinkCreationStateV3 } from "$modules/creationLink/state/linkCreati
 import { LinkCreatedStateV3 } from "$modules/creationLink/state/linkCreationStatesV3/created";
 import { LockStateV3 } from "$modules/creationLink/state/linkCreationStatesV3/lock";
 import type { LinkCreationStoreV3 } from "$modules/creationLink/state/linkCreationStoreV3.svelte";
+import { gateDraftToGateKey } from "$modules/creationLink/utils/gateDraftToGateKey";
 import type { GateKey } from "$lib/generated/cashier_backend/cashier_backend.did";
-import { GateType, type GateDraft } from "$modules/gating/types/gate";
 import { cashierBackendService } from "$modules/links/services/cashierBackend";
 import { LinkStep } from "$modules/links/types/linkStep";
 
 /**
  * State handler for the Preview step in the V3 create-link flow.
  */
-export function gateDraftToGateKey(gateDraft: GateDraft): GateKey {
-  switch (gateDraft.type) {
-    case GateType.PASSWORD:
-      return { Password: gateDraft.password };
-    case GateType.X_FOLLOWING:
-      return { XFollowing: gateDraft.targetHandle };
-    case GateType.X_OWNED_ACCOUNT:
-      return { XOwnedAccount: gateDraft.targetHandle };
-    case GateType.X_LIKED_POST:
-      return { XLikedPost: gateDraft.tweetUrl };
-    case GateType.X_RETWEETED_POST:
-      return { XRetweetedPost: gateDraft.tweetUrl };
-    case GateType.OTP_EMAIL:
-      return { OTPEmail: gateDraft.email };
-    case GateType.OTP_SMS:
-      return { OTPSms: gateDraft.phone };
-  }
-}
-
-// State handler for Preview step in the link creation flow (V3)
 export class PreviewStateV3 implements LinkCreationStateV3 {
   readonly step = LinkStep.PREVIEW;
   #linkStore: LinkCreationStoreV3;
